Removed Source editor
